Moving editor camera controller to editor folder
